# src/Trips.Api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Trips.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5099/"
    }
  ],
  "paths": {
...
    "/trips": {
      "post": {
        "tags": [
          "Trips"
        ],
        "operationId": "CreateTrip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Trips"
        ],
        "operationId": "ListTrips",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TripDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/trips/{id}": {
      "get": {
        "tags": [
          "Trips"
        ],
        "operationId": "GetTrip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Trips"
        ],
        "operationId": "DeleteTrip",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{id}/destination": {
      "patch": {
        "tags": [
          "Trips"
        ],
        "operationId": "UpdateTripDestination",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripDestinationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/participants": {
      "post": {
        "tags": [
          "Participants"
        ],
        "operationId": "AddParticipant",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddParticipantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/participants/{pid}/prefs": {
      "patch": {
        "tags": [
          "Participants"
        ],
        "operationId": "UpdateParticipantPrefs",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreferencesDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/trips/{tripId}/participants/{pid}": {
      "delete": {
        "tags": [
          "Participants"
        ],
        "operationId": "RemoveParticipant",
        "parameters": [
          {
            "name": "tripId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddParticipantRequest": {
        "required": [
          "displayName",
          "homeAddress",
          "homeLongitude",
          "homeLatitude",
          "hasCar",
          "seats",
          "preferences"
        ],
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "homeAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "homeLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "homeLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "hasCar": {
            "type": "boolean"
          },
          "seats": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "preferences": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PreferencesDto"
              }
            ]
          }
        }
      },
      "CandidateNodeDto": {
        "required": [
          "id",
          "participantId",
          "kind",
          "longitude",
          "latitude",
          "walkMins",
          "ptMins",
          "externalId",
          "displayName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "participantId": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/CandidateNodeKindDto"
          },
          "longitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "latitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "walkMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ptMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "externalId": {
            "type": [
              "null",
              "string"
            ]
          },
          "displayName": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CandidateNodeKindDto": {
        "enum": [
          "home",
          "trainStation",
          "busStop",
          "wharf",
          "lightRailStop"
        ]
      },
...
      "CreateTripRequest": {
        "required": [
          "name",
          "destinationName",
          "destinationLongitude",
          "destinationLatitude",
          "departAt",
          "arrivalWindowEarliest",
          "arrivalWindowLatest"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "destinationName": {
            "type": "string"
          },
          "destinationLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "destinationLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "departAt": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowEarliest": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowLatest": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ParticipantDto": {
        "required": [
          "id",
          "tripId",
          "displayName",
          "homeLongitude",
          "homeLatitude",
          "hasCar",
          "seats",
          "preferences"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "homeLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "homeLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "hasCar": {
            "type": "boolean"
          },
          "seats": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesDto"
          }
        }
      },
      "ParticipantWithNodesDto": {
        "required": [
          "id",
          "tripId",
          "displayName",
          "homeLongitude",
          "homeLatitude",
          "hasCar",
          "seats",
          "preferences",
          "candidateNodes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tripId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string"
          },
          "homeLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "homeLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "hasCar": {
            "type": "boolean"
          },
          "seats": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesDto"
          },
          "candidateNodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateNodeDto"
            }
          }
        }
      },
      "PreferencesDto": {
        "required": [
          "walkBudgetMins",
          "detourToleranceMins",
          "fairnessWeight"
        ],
        "type": "object",
        "properties": {
          "walkBudgetMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detourToleranceMins": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fairnessWeight": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
      "TripDetailDto": {
        "required": [
          "id",
          "name",
          "destinationName",
          "destinationLongitude",
          "destinationLatitude",
          "departAt",
          "arrivalWindowEarliest",
          "arrivalWindowLatest",
          "ownerId",
          "createdAt",
          "lockedSolutionId",
          "participants"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "destinationName": {
            "type": "string"
          },
          "destinationLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "destinationLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "departAt": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowEarliest": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowLatest": {
            "type": "string",
            "format": "date-time"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockedSolutionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantWithNodesDto"
            }
          }
        }
      },
      "TripDto": {
        "required": [
          "id",
          "name",
          "destinationName",
          "destinationLongitude",
          "destinationLatitude",
          "departAt",
          "arrivalWindowEarliest",
          "arrivalWindowLatest",
          "ownerId",
          "createdAt",
          "lockedSolutionId",
          "participantCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "destinationName": {
            "type": "string"
          },
          "destinationLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "destinationLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "departAt": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowEarliest": {
            "type": "string",
            "format": "date-time"
          },
          "arrivalWindowLatest": {
            "type": "string",
            "format": "date-time"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lockedSolutionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "participantCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "UpdateTripDestinationRequest": {
        "required": [
          "destinationName",
          "destinationLongitude",
          "destinationLatitude"
        ],
        "type": "object",
        "properties": {
          "destinationName": {
            "type": "string"
          },
          "destinationLongitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "destinationLatitude": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?(?:[eE][+-]?\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          }
        }
      },
...
    }
  },
...
}